requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
litellm
pyyaml
sqlalchemy
asyncpg
aiohttp
Pillow
